+++ app/models/device_token.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class DeviceToken < ApplicationRecord
+  validates :token, presence: true, uniqueness: true
+end